fix:테이블 수정으로 인한 페키지 dto변경
