--- src/main/java/kr/co/iei/comm/controller/CommController.java
@@ -1,17 +1,0 @@
-package kr.co.iei.comm.controller;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import kr.co.iei.comm.model.service.CommService;
-
-@CrossOrigin("*")
-@RestController
-@RequestMapping(value="/comm")
-public class CommController {
-	@Autowired
-	private CommService commService;
-
-}
