--- src/main/java/kr/co/iei/comm/model/dao/PlanCommDao.java
@@ -1,8 +1,0 @@
-package kr.co.iei.comm.model.dao;
-
-import org.apache.ibatis.annotations.Mapper;
-
-@Mapper
-public interface PlanCommDao {
-
-}
