--- src/main/java/kr/co/iei/comm/model/service/CommService.java
@@ -1,12 +1,0 @@
-package kr.co.iei.comm.model.service;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import kr.co.iei.comm.model.dao.CommDao;
-
-@Service
-public class CommService {
-	@Autowired
- private CommDao commDao;
-}
