+++ src/main/java/kr/co/iei/review/model/dao/CommDao.java
@@ -3,6 +3,6 @@ package kr.co.iei.review.model.dao;
 import org.apache.ibatis.annotations.Mapper;
 
 @Mapper
-public interface PlanReviewDao {
+public interface CommDao {
 
 }
