+++ src/main/java/kr/co/iei/review/model/service/CommService.java
@@ -3,10 +3,12 @@ package kr.co.iei.review.model.service;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import kr.co.iei.review.model.dao.PlanReviewDao;
+import kr.co.iei.review.model.dao.CommDao;
+
+
 
 @Service
-public class PlanReviewService {
- @Autowired
- private PlanReviewDao planReviewDao;
+public class CommService {
+	@Autowired
+ private CommDao commDao;
 }
